pyproject.toml
[project]
name = "crypto-daytrading-arena"
version = "0.1.0"
description = "Multi-agent crypto daytrading arena powered by Calfkit"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "calfkit>=0.1.10",
    "httpx>=0.27",
    "plotext>=5.3.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.0.0",
    "sympy>=1.13",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s %(levelname)-5s %(name)s — %(message)s"
log_cli_date_format = "%H:%M:%S"
